+++ Changelog.md
@@ -2,8 +2,8 @@
 
 ## 3.x Series
 
-## Unreleased
-##### 2023-XX-YY
+## 3.6.2
+##### 2023-03-12
 
 - Fixed the shipping adjustment's undesired JSON->string conversion when cloning the cart adjustment to the order
 - Fixed the missing locking of shipping adjustments on orders